+++ collectivo/collectivo/composables/directus.ts
@@ -3,13 +3,13 @@ export const useDirectus = () => {
   return $directus;
 };
 
-export function requireAuth() {
+export function requireAuth(force: boolean = false) {
   const directus = useDirectus();
   const user = useCollectivoUser();
   const runtimeConfig = useRuntimeConfig();
 
   // If user is authenticated, do nothing
-  if (user.value.isAuthenticated === true) return;
+  if (user.value.isAuthenticated === true && !force) return;
 
   // If user is not authenticated, log out of directus and redirect to keycloak
   directus.logout();
