--- collectivo/collectivo/composables/forms.ts
@@ -1,2 +1,0 @@
-export const useCollectivoForms = () =>
-  useState<{ [key: string]: CollectivoForm }>("collectivo_forms", () => ({}));
