--- collectivo/extensions/memberships/app.config.ts
@@ -1,12 +1,0 @@
-import { IdentificationIcon } from "@heroicons/vue/24/solid";
-
-export default defineAppConfig({
-  mainMenuItems: {
-    membership: {
-      label: "Membership",
-      link: "/membership",
-      icon: IdentificationIcon,
-      order: 10,
-    },
-  },
-});
