--- collectivo/extensions/memberships/pages/membership/index.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-setCollectivoTitle("Membership");
-</script>
-
-<template>
-  <div>MEMBERSHIP SECTION</div>
-</template>
